create a dedicated component for the video item + change the link + update the behavior
